Add controller [#159837162]
-Add controller folder and questionsControllers.js file to
 make the routes readable, now we can better understand the code using
 the single responsibility principle
